+++ benchmarks/power_distribution/power_distributor.py
@@ -137,10 +137,9 @@ async def run_test(  # pylint: disable=too-many-locals
 
 async def run() -> None:
     """Create microgrid api and run tests."""
-    # pylint: disable=protected-access
-
     await microgrid.initialize(
-        HOST, PORT, ResamplerConfig(resampling_period=timedelta(seconds=1.0))
+        "grpc://microgrid.sandbox.api.frequenz.io:62060",
+        ResamplerConfig(resampling_period=timedelta(seconds=1.0)),
     )
 
     all_batteries: set[Component] = connection_manager.get().component_graph.components(