+++ examples/battery_pool.py
@@ -22,9 +22,11 @@ async def main() -> None:
     logging.basicConfig(
         level=logging.DEBUG, format="%(asctime)s %(name)s %(levelname)s:%(message)s"
     )
+    # Reduce undesired noise from hpack
+    logging.getLogger("hpack.hpack").setLevel(logging.INFO)
+
     await microgrid.initialize(
-        host=HOST,
-        port=PORT,
+        "grpc://microgrid.sandbox.api.frequenz.io:62060",
         resampler_config=ResamplerConfig(resampling_period=timedelta(seconds=1.0)),
     )