+++ src/frequenz/sdk/actor/power_distributing/_component_status/_battery_status_tracker.py
@@ -21,19 +21,16 @@ import math
 from dataclasses import dataclass
 from datetime import datetime, timedelta, timezone
 
-# pylint: disable=no-name-in-module
-from frequenz.api.microgrid.battery_pb2 import ComponentState as BatteryComponentState
-from frequenz.api.microgrid.battery_pb2 import RelayState as BatteryRelayState
-from frequenz.api.microgrid.common_pb2 import ErrorLevel
-from frequenz.api.microgrid.inverter_pb2 import ComponentState as InverterComponentState
-
-# pylint: enable=no-name-in-module
 from frequenz.channels import Receiver, Sender, select, selected_from
 from frequenz.channels.timer import SkipMissedAndDrift, Timer
 from frequenz.client.microgrid import (
+    BatteryComponentState,
     BatteryData,
+    BatteryRelayState,
     ComponentCategory,
     ComponentData,
+    ErrorLevel,
+    InverterComponentState,
     InverterData,
 )
 from typing_extensions import override
@@ -72,29 +69,27 @@ class BatteryStatusTracker(ComponentStatusTracker, BackgroundService):
     Status updates are sent out only when there is a status change.
     """
 
-    _battery_valid_relay: set[BatteryRelayState.ValueType] = {
-        BatteryRelayState.RELAY_STATE_CLOSED
-    }
+    _battery_valid_relay: set[BatteryRelayState] = {BatteryRelayState.CLOSED}
     """The list of valid relay states of a battery.
 
     A working battery in any other battery relay state will be reported as failing.
     """
 
-    _battery_valid_state: set[BatteryComponentState.ValueType] = {
-        BatteryComponentState.COMPONENT_STATE_IDLE,
-        BatteryComponentState.COMPONENT_STATE_CHARGING,
-        BatteryComponentState.COMPONENT_STATE_DISCHARGING,
+    _battery_valid_state: set[BatteryComponentState] = {
+        BatteryComponentState.IDLE,
+        BatteryComponentState.CHARGING,
+        BatteryComponentState.DISCHARGING,
     }
     """The list of valid states of a battery.
 
     A working battery in any other battery state will be reported as failing.
     """
 
-    _inverter_valid_state: set[InverterComponentState.ValueType] = {
-        InverterComponentState.COMPONENT_STATE_STANDBY,
-        InverterComponentState.COMPONENT_STATE_IDLE,
-        InverterComponentState.COMPONENT_STATE_CHARGING,
-        InverterComponentState.COMPONENT_STATE_DISCHARGING,
+    _inverter_valid_state: set[InverterComponentState] = {
+        InverterComponentState.STANDBY,
+        InverterComponentState.IDLE,
+        InverterComponentState.CHARGING,
+        InverterComponentState.DISCHARGING,
     }
     """The list of valid states of an inverter.
 
@@ -370,11 +365,11 @@ class BatteryStatusTracker(ComponentStatusTracker, BackgroundService):
         Returns:
             True if message has no critical error, False otherwise.
         """
-        critical = ErrorLevel.ERROR_LEVEL_CRITICAL
-        # pylint: disable=protected-access
-        critical_err = next((err for err in msg._errors if err.level == critical), None)
+        critical = ErrorLevel.CRITICAL
+        critical_err = next((err for err in msg.errors if err.level == critical), None)
         if critical_err is not None:
-            if self._last_status == ComponentStatusEnum.WORKING:
+            last_status = self._last_status  # pylint: disable=protected-access
+            if last_status == ComponentStatusEnum.WORKING:
                 _logger.warning(
                     "Component %d has critical error: %s",
                     msg.component_id,
@@ -394,15 +389,16 @@ class BatteryStatusTracker(ComponentStatusTracker, BackgroundService):
         """
         # Component state is not exposed to the user.
         # pylint: disable=protected-access
-        state = msg._component_state
+        state = msg.component_state
         if state not in BatteryStatusTracker._inverter_valid_state:
             if self._last_status == ComponentStatusEnum.WORKING:
                 _logger.warning(
                     "Inverter %d has invalid state: %s",
                     msg.component_id,
-                    InverterComponentState.Name(state),
+                    state.name,
                 )
             return False
+        # pylint: enable=protected-access
         return True
 
     def _is_battery_state_correct(self, msg: BatteryData) -> bool:
@@ -416,28 +412,28 @@ class BatteryStatusTracker(ComponentStatusTracker, BackgroundService):
         """
         # Component state is not exposed to the user.
         # pylint: disable=protected-access
-        state = msg._component_state
+        state = msg.component_state
         if state not in BatteryStatusTracker._battery_valid_state:
             if self._last_status == ComponentStatusEnum.WORKING:
                 _logger.warning(
                     "Battery %d has invalid state: %s",
                     self.battery_id,
-                    BatteryComponentState.Name(state),
+                    state.name,
                 )
             return False
 
         # Component state is not exposed to the user.
-        # pylint: disable=protected-access
-        relay_state = msg._relay_state
+        relay_state = msg.relay_state
         if relay_state not in BatteryStatusTracker._battery_valid_relay:
             if self._last_status == ComponentStatusEnum.WORKING:
                 _logger.warning(
                     "Battery %d has invalid relay state: %s",
                     self.battery_id,
-                    BatteryRelayState.Name(relay_state),
+                    relay_state.name,
                 )
             return False
         return True
+        # pylint: enable=protected-access
 
     def _is_timestamp_outdated(self, timestamp: datetime) -> bool:
         """Return if timestamp is to old.