+++ src/frequenz/sdk/timeseries/logical_meter/_logical_meter.py
@@ -36,8 +36,7 @@ class LogicalMeter:
         from frequenz.sdk.timeseries import ResamplerConfig
 
         await microgrid.initialize(
-            "127.0.0.1",
-            50051,
+            "grpc://127.0.0.1:50051",
             ResamplerConfig(resampling_period=timedelta(seconds=1))
         )